pyproject.toml
[project]
name = "testfordb"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "asyncmy>=0.2.11",
    "cryptography>=46.0.6",
    "elasticsearch[async]>=8,<9",
    "fastapi[standard]>=0.135.3",
    "jieba>=0.42.1",
    "langchain>=1.2.15",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "loguru>=0.7.3",
    "omegaconf>=2.3.0",
    "paddle>=1.3.0",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.17.1",
    "sqlalchemy>=2.0.49",
]
